Add files via upload
LED Blinking Code for Arduino
Description: This code is designed to make an LED blink with specific timing intervals using an Arduino board. The LED connected to pin 19 will turn on for 1 second, then turn off for 0.3 seconds, creating a repeating blink pattern.

Code Functionality
LED Pin: The LED is connected to digital pin 19, which is defined as led in the code.
Setup: The setup() function initializes pin 19 as an output, preparing it for controlling the LED.
Loop: In the loop() function:
The LED is turned on (HIGH) for 1 second (delay(1000)).
The LED is turned off (LOW) for 0.3 seconds (delay(300)).
This pattern repeats continuously.